# Improvements to the "Sum" SIMD algorithm