--- a/test/Microsoft.ML.CpuMath.PerformanceTests/CpuMathNativeUtils.cs
+++ b/test/Microsoft.ML.CpuMath.PerformanceTests/CpuMathNativeUtils.cs
@@ -50,8 +50,8 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         [DllImport("CpuMathNative", EntryPoint = "MulElementWiseU"), SuppressUnmanagedCodeSecurity]
         internal static extern unsafe void MulElementWiseU(/*_In_ const*/ float* ps1, /*_In_ const*/ float* ps2, /*_Inout_*/ float* pd, int c);
 
-        [DllImport("CpuMathNative", EntryPoint = "SumU"), SuppressUnmanagedCodeSecurity]
-        internal static extern unsafe float SumU(/*const*/ float* ps, int c);
+        [DllImport("CpuMathNative", EntryPoint = "Sum"), SuppressUnmanagedCodeSecurity]
+        internal static extern unsafe float Sum(/*const*/ float* pValues, int length);
 
         [DllImport("CpuMathNative", EntryPoint = "SumSqU"), SuppressUnmanagedCodeSecurity]
         internal static extern unsafe float SumSqU(/*const*/ float* ps, int c);