--- a/test/Microsoft.ML.TestFramework/DataPipe/TestDataPipeBase.cs
+++ b/test/Microsoft.ML.TestFramework/DataPipe/TestDataPipeBase.cs
@@ -137,7 +137,7 @@ namespace Microsoft.ML.Runtime.RunTests
                 // This in turn means that the schema of loaded transformer matches for 
                 // Transform and GetOutputSchema calls.
                 CheckSameSchemas(scoredTrain.Schema, scoredTrain2.Schema);
-                CheckSameValues(scoredTrain, scoredTrain2);
+                CheckSameValues(scoredTrain, scoredTrain2, exactDoubles: false);
             };
 
             checkOnData(validFitInput);
@@ -1092,7 +1092,7 @@ namespace Microsoft.ML.Runtime.RunTests
                         if (exactDoubles)
                             return GetComparerOne<Double>(r1, r2, col, (x, y) => FloatUtils.GetBits(x) == FloatUtils.GetBits(y));
                         else
-                            return GetComparerOne<Double>(r1, r2, col, EqualWithEps);
+                            return GetComparerOne<Double>(r1, r2, col, EqualWithEpsDouble);
                     case DataKind.Text:
                         return GetComparerOne<ReadOnlyMemory<char>>(r1, r2, col, (a ,b) => a.Span.SequenceEqual(b.Span));
                     case DataKind.Bool:
@@ -1133,12 +1133,15 @@ namespace Microsoft.ML.Runtime.RunTests
                     case DataKind.U8:
                         return GetComparerVec<ulong>(r1, r2, col, size, (x, y) => x == y);
                     case DataKind.R4:
-                        return GetComparerVec<Single>(r1, r2, col, size, (x, y) => FloatUtils.GetBits(x) == FloatUtils.GetBits(y));
+                        if (exactDoubles)
+                            return GetComparerVec<Single>(r1, r2, col, size, (x, y) => FloatUtils.GetBits(x) == FloatUtils.GetBits(y));
+                        else
+                            return GetComparerVec<Single>(r1, r2, col, size, EqualWithEpsSingle);
                     case DataKind.R8:
                         if (exactDoubles)
                             return GetComparerVec<Double>(r1, r2, col, size, (x, y) => FloatUtils.GetBits(x) == FloatUtils.GetBits(y));
                         else
-                            return GetComparerVec<Double>(r1, r2, col, size, EqualWithEps);
+                            return GetComparerVec<Double>(r1, r2, col, size, EqualWithEpsDouble);
                     case DataKind.Text:
                         return GetComparerVec<ReadOnlyMemory<char>>(r1, r2, col, size, (a, b) => a.Span.SequenceEqual(b.Span));
                     case DataKind.Bool:
@@ -1176,10 +1179,18 @@ namespace Microsoft.ML.Runtime.RunTests
 
         private const Double DoubleEps = 1e-9;
 
-        private static bool EqualWithEps(Double x, Double y)
+        private static bool EqualWithEpsDouble(Double x, Double y)
         {
             // bitwise comparison is needed because Abs(Inf-Inf) and Abs(NaN-NaN) are not 0s.
             return FloatUtils.GetBits(x) == FloatUtils.GetBits(y) || Math.Abs(x - y) < DoubleEps;
+        }
+
+        private const float SingleEps = 1e-6f;
+
+        private static bool EqualWithEpsSingle(float x, float y)
+        {
+            // bitwise comparison is needed because Abs(Inf-Inf) and Abs(NaN-NaN) are not 0s.
+            return FloatUtils.GetBits(x) == FloatUtils.GetBits(y) || Math.Abs(x - y) < SingleEps;
         }
 
         protected Func<bool> GetComparerOne<T>(IRow r1, IRow r2, int col, Func<T, T, bool> fn)