Does some cleanup so that we have a single "Sum" algorithm (rather than one for aligned and one for unaligned inputs).

For inputs with fewer elements than can fit in the `Vector` type, it falls back to scalar code.
For inputs that are not naturally aligned (the alignment is not a multiple of 4), it does exclusively unaligned loads
For all other inputs, it will do at most two unaligned loads (one each for any leading/trailing unaligned elements) and all other loads will be aligned.