@tannergooding 2018-10-02
This is a simple example of https://github.com/dotnet/machinelearning/issues/836