@tannergooding 2018-10-02
Results in some minor perf improvements for the `Microsoft.ML.CpuMath.PerformanceTests`.

Before:

|      Method |     Mean |     Error |    StdDev |
|------------ |---------:|----------:|----------:|
|    Avx.SumU | 149.6 us | 0.9997 us | 0.9351 us |
| Native.SumU | 262.2 us |  2.064 us |  1.931 us |
|    Sse.SumU | 263.1 us |  1.504 us |  1.407 us |

After:

|    Method  |     Mean |    Error |    StdDev |
|----------- |---------:|---------:|----------:|
|    Avx.Sum | 129.7 us | 1.103 us | 0.9775 us |
| Native.Sum | 261.2 us | 1.013 us | 0.8983 us |
|    Sse.Sum | 255.3 us | 1.521 us |  1.422 us |