@tannergooding 2018-10-02
This can serve as the basis for the other algorithms as well. Generally the only tweaking that needs to happen is when dealing with leading/trailing elements, where you may need additional masking/etc to get the elements lined up correctly. For example, `Scale` requires you to ensure the masked out elements are the original value, rather than zero (which requires a couple additional instructions).