@tannergooding 2018-10-02
Test failure is for `NormalizerTests.LpGcNormAndWhiteningWorkout` and is due to a baseline diff for `NormalizerTests.LpGcNormAndWhiteningWorkout` where the last couple digits of the result may differ due to different indices being summed together depending on the alignment of the input data (for floating point: `a + b + c` can produce a different result than `a + c + b`).