@tannergooding 2018-10-02
Baseline
```
#@ TextLoader{
#@   sep=tab
#@   col=lpnorm:R4:0-10
#@   col=gcnorm:R4:11-21
#@   col=whitened:R4:22-32
#@ }
-0.686319232	0.192169383	-0.152238086	0.03493989	0.346903175	0.09483684	-0.132272437	-0.124785319	-0.5315855	-0.0973325446	0.114802495	-0.626524031	0.289601743	-0.0695612058	0.125636056	0.4509648	0.188099176	-0.0487401523	-0.04093227	-0.465160966	-0.0123033375	0.208920211	-2.604605	0.829638362	-0.5992434	0.19860521	1.33247662	0.369197041	-0.5760094	-0.5490271	-1.94509208	-0.393351972	0.507488966
-0.20306389	-0.1231699	-0.039946992	0.183090389	-0.3328916	0.279628932	-0.0066578323	0.432759076	-0.0798939839	-0.1664458	-0.7057302	-0.137441739	-0.055349838	0.0301625486	0.259335726	-0.270841062	0.3585301	0.0643675	0.5158729	-0.0108833946	-0.09981628	-0.653936446	-0.5923902	-0.324390084	-0.114805378	0.6855182	-1.055579	0.8767955	-0.0392023772	1.21807373	-0.160801888	-0.47570774	-2.22817
-0.268398017	-0.28734377	0.571529865	0.006315247	-0.246294647	-0.445224941	-0.344181	-0.20524554	0.284186125	-0.116832078	-0.06946772	-0.176903129	-0.19703348	0.715542555	0.114987023	-0.153417692	-0.3647864	-0.257424533	-0.109801926	0.410232216	-0.0158602837	0.0344656035	-0.9132714	-0.911281645	1.814283	0.07471426	-0.8969923	-1.44387519	-1.19571114	-0.6542767	0.887983143	-0.4604767	-0.17543222
0.117021732	0.438831449	-0.100304335	0.125380427	-0.413755417	0.0794076	0.133739114	-0.397038	-0.497342378	-0.2632989	0.313451052	0.160775661	0.485780418	-0.0587080531	0.169217348	-0.3752711	0.122788094	0.17765902	-0.358387738	-0.459687948	-0.223320842	0.3591552	0.236966148	1.004758	-0.233154371	0.3862052	-1.02724624	0.240614042	0.299898773	-1.03102541	-1.13852251	-0.6675951	0.766793966
```

Diff:
```
#@ TextLoader{
#@   sep=tab
#@   col=lpnorm:R4:0-10
#@   col=gcnorm:R4:11-21
#@   col=whitened:R4:22-32
#@ }
-0.686319232	0.192169383	-0.152238086	0.03493989	0.346903175	0.09483684	-0.132272437	-0.124785319	-0.5315855	-0.0973325446	0.114802495	-0.626524031	0.289601743	-0.0695612058	0.125636056	0.4509648	0.188099176	-0.0487401523	-0.04093227	-0.465160966	-0.0123033375	0.208920211	-2.604605	0.829638362	-0.5992434	0.19860521	1.33247662	0.369197041	-0.5760094	-0.5490271	-1.94509208	-0.393351972	0.507488966
-0.20306389	-0.1231699	-0.039946992	0.183090389	-0.3328916	0.279628932	-0.0066578323	0.432759076	-0.0798939839	-0.1664458	-0.7057302	-0.137441739	-0.055349838	0.0301625486	0.259335726	-0.270841062	0.3585301	0.0643675	0.5158729	-0.0108833946	-0.09981628	-0.653936446	-0.5923902	-0.324390084	-0.114805378	0.6855182	-1.055579	0.8767955	-0.0392023772	1.21807373	-0.160801888	-0.47570774	-2.22817
-0.268398017	-0.28734377	0.571529865	0.006315247	-0.246294647	-0.445224941	-0.344181	-0.20524554	0.284186125	-0.116832078	-0.06946772	-0.1769031	-0.19703348	0.715542555	0.114987031	-0.153417677	-0.3647864	-0.257424533	-0.109801918	0.410232216	-0.0158602744	0.0344656147	-0.9132714	-0.911281645	1.814283	0.07471426	-0.8969923	-1.44387519	-1.19571114	-0.6542767	0.887983143	-0.4604767	-0.17543222
0.117021732	0.438831449	-0.100304335	0.125380427	-0.413755417	0.0794076	0.133739114	-0.397038	-0.497342378	-0.2632989	0.313451052	0.160775676	0.485780418	-0.0587080531	0.169217363	-0.3752711	0.122788109	0.177659035	-0.358387738	-0.459687948	-0.223320842	0.3591552	0.236966148	1.004758	-0.233154371	0.3862052	-1.02724624	0.240614042	0.299898773	-1.03102541	-1.13852251	-0.6675951	0.766793966
```