@tannergooding 2018-10-23
CC. @eerhardt, @Anipik for review.