@tannergooding 2018-10-24
Rebased to resolve conflicts.